simple error handling

+++ gameDevTycoonFixer.py
@@ -28,19 +28,26 @@ while nothingFixed:
 	if output == "" or (command in output and not searchForCommand in output):
 		print("game not running, please start 'Game Dev Tycoon' ...")
 	elif searchForCommand in output:
-		outputParts = output.split(" ")
+		try:
+			outputParts = output.split(" ")
 		
-		for cmd in outputParts:
-			if cmd.startswith(searchForCommand):
-				workingPath = cmd.replace(searchForCommand, "")
+			for cmd in outputParts:
+				if cmd.startswith(searchForCommand):
+					workingPath = cmd.replace(searchForCommand, "")
 				
-				source = workingPath + "/images/superb/level2Desk.png"
-				dest = workingPath + "/images/superb/level2desk.png"
-				os.rename(source, dest)
-				
-				print("found what we were searching for and fixed it ;)")
-				nothingFixed = False
-				break
+					source = workingPath + "/images/superb/level2Desk.png"
+					dest = workingPath + "/images/superb/level2desk.png"
+					
+					if os.path.isfile(source):
+						os.rename(source, dest)
+						
+						print("found what we were searching for and fixed it ;)")
+						nothingFixed = False
+						break
+					else:
+						print("game is not ready yet")
+		except Exception as ex:
+			print(ex)
 	else:
 		print("Found unknown output:\n" + output)
